enable compiler warnings

+++ source/options.hpp
@@ -14,8 +14,8 @@ string wrap_help(const string& option, const string& text, const unsigned short 
 
 bool output_directory_exists(const string& output_file);
 
-bool validate_int(const char* optarg, int& value, const int min_value = LONG_MIN, const int max_value = LONG_MAX);
-bool validate_int(const char* optarg, unsigned int& value, const unsigned int min_value = 0, const unsigned int max_value = LONG_MAX);
+bool validate_int(const char* optarg, int& value, const int min_value = INT_MIN, const int max_value = INT_MAX);
+bool validate_int(const char* optarg, unsigned int& value, const unsigned int min_value = 0, const unsigned int max_value = INT_MAX);
 
 bool validate_float(const char* optarg, float& value, const float min_value = FLT_MIN, const float max_value = FLT_MAX);
